final_pjt_back/requirements.txt
asgiref==3.3.4
asttokens==2.0.8
backcall==0.2.0
colorama==0.4.4
decorator==5.0.7
Django==3.2
django-cors-headers==4.0.0
django-extensions==3.1.3
django-seed==0.2.2
djangorestframework==3.12.4
executing==0.10.0
Faker==8.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.6
parso==0.8.2
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.18
pure-eval==0.2.2
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
stack-data==0.4.0
text-unidecode==1.3
traitlets==5.0.5
tzdata==2022.2
wcwidth==0.2.5
